+++ wxc_sdk/telephony/__init__.py
@@ -23,6 +23,7 @@ from .organisation_vm import OrganisationVoicemailSettingsAPI
 from .paging import PagingApi
 from .pnc import PrivateNetworkConnectApi
 from .prem_pstn import PremisePstnApi
+from .virtual_line import VirtualLinesApi
 from .vm_rules import VoicemailRulesApi
 from .voice_messaging import VoiceMessagingApi
 from .voicemail_groups import VoicemailGroupsApi
@@ -447,6 +448,7 @@ class TelephonyApi(ApiChild, base='telephony/config'):
     prem_pstn: PremisePstnApi
     pnc: PrivateNetworkConnectApi
     schedules: ScheduleApi
+    virtual_lines: VirtualLinesApi
     # location voicemail groups
     voicemail_groups: VoicemailGroupsApi
     voicemail_rules: VoicemailRulesApi
@@ -473,6 +475,7 @@ class TelephonyApi(ApiChild, base='telephony/config'):
         self.pnc = PrivateNetworkConnectApi(session=session)
         self.prem_pstn = PremisePstnApi(session=session)
         self.schedules = ScheduleApi(session=session, base=ScheduleApiBase.locations)
+        self.virtual_lines = VirtualLinesApi(session=session)
         self.voicemail_groups = VoicemailGroupsApi(session=session)
         self.voicemail_rules = VoicemailRulesApi(session=session)
         self.voice_messaging = VoiceMessagingApi(session=session)